POC/pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "av>=16.1.0",
    "imagehash>=4.3.2",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pywavelets>=1.9.0",
    "scipy>=1.17.1",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
